feat: runAuger, fixed num rotations in runHopper.

+++ src/main.cpp
@@ -300,8 +300,8 @@ Run the hopper once to the number of revolutions needed to dispense the whole vo
 void runHopper(AccelStepper& stepper, Encoder& encoder) {
     const int encoder_ppr = 20*4;           // 20 ppr encoder, res of 4/ppr using encoder.h
     const int driver_ppr = 400;             // set on stepper controller
-    const int num_revs = 1/5;
-    int steps = driver_ppr * num_revs;
+    const float num_revs = 1/5;
+    long steps = driver_ppr * num_revs;
     long finalEncPos = encoder_ppr * num_revs;
     long encPos = 0;
